requirements.txt
openai
python-dotenv
@ai-sdk/xai
